refactor: Decouple HTTP caching via factory interface
Introduce ICachingHandlerFactory to abstract caching handler creation and move CacheCow-based caching to new Api.Rest.Caching project.

Update RestClient and builder to use the new factory, removing direct CacheCow dependencies.
